move runner id generator into core
