+++ packages/core/src/v3/isomorphic/friendlyId.ts
@@ -90,3 +90,25 @@ export const RunId = new IdUtil("run");
 export const SnapshotId = new IdUtil("snapshot");
 export const WaitpointId = new IdUtil("waitpoint");
 export const BatchId = new IdUtil("batch");
+
+export class IdGenerator {
+  private alphabet: string;
+  private length: number;
+  private prefix: string;
+
+  constructor({ alphabet, length, prefix }: { alphabet: string; length: number; prefix: string }) {
+    this.alphabet = alphabet;
+    this.length = length;
+    this.prefix = prefix;
+  }
+
+  generate(): string {
+    return `${this.prefix}${customAlphabet(this.alphabet, this.length)()}`;
+  }
+}
+
+export const RunnerId = new IdGenerator({
+  alphabet: "123456789abcdefghijkmnopqrstuvwxyz",
+  length: 20,
+  prefix: "runner_",
+});
